dom/file/test/test_lockedfile_lifetimes.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    for each (let fileStorage in fileStorages) {
      let request = getFileHandle(fileStorage.key, "test.txt");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      let fileHandle = event.target.result;
      fileHandle.onerror = errorHandler;

      let lockedFile = fileHandle.open();
      continueToNextStep();
      yield undefined;

      try {
        lockedFile.getMetadata({ size: true });
        ok(false, "Should have thrown!");
      }
      catch (e) {
        ok(e instanceof DOMException, "Got exception.");
        is(e.name, "LockedFileInactiveError", "Good error.");
        is(e.code, 0, "Good error code.");
      }
    }

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
